wag up

+++ app/src/main/java/com/tttrtclive/live/LocalConfig.java
@@ -3,7 +3,7 @@ package com.tttrtclive.live;
 public class LocalConfig {
 
     public static long mLocalUserID;
-    public static long mLocalRole;
+    public static int mLocalRole;
 
     public static int mLocalHeight;
     public static int mLocalWidth;
